Removed IOErrors from FortranFile.
IOError should only be raised if there is a real problem actually
reading data from disk, not just when the data read isn't what was
expected. FortranFile now has two new exceptions NoMoreRecords and
IntegrityError derived from Exception to indicated when you've reached
the end of the file or if the file appears to be corrupt, respectively.

Successful reads still have the exact same API, but any code that relies
on catching IOErrors will need to be updated.