Deleted the custom endian settings.
I wasn't using it, plus for some reason it made everything slower.
Profiling put it in the _get_endian method a large fraction of the time.
I don't understand why, but after removing all of that now most of the
time is spent actually reading the file. Most of the time is spent split
between reading bytes for the record sizes or in numpy.readstring for
reading the actual records.